chore: checkpoint before using codex. implement inventory domain with clean architecture

+++ app/Domains/Inventory/Application/DTOs/Inventario/InventoryOutput.php
@@ -9,7 +9,7 @@ class InventoryOutput
         public readonly string $name,
         public readonly string $unit,
         public readonly float $currentQuantity,
-        public readonly int $costPrice,
+        public readonly float $costPrice,
         public readonly int $restaurantId,
         public readonly int $supplierId
     ) {}
